Major update on module related stuff
