+++ components/module/AnnouncementItem.tsx
@@ -9,15 +9,14 @@ import {
   useToast,
 } from '@chakra-ui/react'
 import { useSession } from 'next-auth/client'
-import router from 'next/router'
-import React, { useState } from 'react'
+import { useRef, useState } from 'react'
 import { FaEdit } from 'react-icons/fa'
 import { ImCancelCircle } from 'react-icons/im'
 import { RiDeleteBin5Line } from 'react-icons/ri'
 import { renderMdToHtml, showCurrentDate } from '../common/Util'
 import DisplayName from '../profile/DisplayName'
 import { useUserId } from '../store/user'
-import { Announcement, deleteAnnouncement } from './DashboardAPI'
+import { Announcement, deleteAnnouncement } from './ModuleAPI'
 import NewAnnouncement from './NewAnnouncement'
 
 export default function AnnouncementItem({
@@ -31,41 +30,38 @@ export default function AnnouncementItem({
   //Alert Dialog
   const [isOpen, setIsOpen] = useState(false)
   const onClose = (): void => setIsOpen(false)
-  const cancelRef = React.useRef()
+  const cancelRef = useRef()
 
-  //Toast
   const toast = useToast()
   const userId = useUserId()
 
   return (
     <>
       <div className="shadow-md w-full border border-gray-100 bg-white dark:bg-gray-700 dark:border-gray-800 relative overflow-hidden p-2 md:p-6 mb-4 mt-2">
-        <>
-          {editing ? (
-            <div className="my-2">
-              <NewAnnouncement
-                label="Edit Announcement"
-                currentAnnouncement={announcement}
-                setEditing={setEditing}
-              />
+        {editing ? (
+          <div className="my-2">
+            <NewAnnouncement
+              label="Edit Announcement"
+              currentAnnouncement={announcement}
+              setEditing={setEditing}
+            />
+          </div>
+        ) : (
+          <div className="flex flex-col">
+            <div className="text-lg font-semibold text-indigo-500">{announcement.title}</div>
+            <div className="text-xs font-thin">
+              <span>
+                <DisplayName author_id={announcement.author_id} />{' '}
+              </span>
+              <span>on {showCurrentDate(announcement.created_date)}</span>
             </div>
-          ) : (
-            <div className="flex flex-col">
-              <div className="text-lg font-semibold text-indigo-500">{announcement.title}</div>
-              <div className="text-xs font-thin">
-                <span>
-                  <DisplayName author_id={announcement.author_id} />{' '}
-                </span>
-                <span>on {showCurrentDate(announcement.created_date)}</span>
-              </div>
-              <div
-                className="py-2"
-                dangerouslySetInnerHTML={{
-                  __html: renderMdToHtml(announcement.content),
-                }}></div>
-            </div>
-          )}
-        </>
+            <div
+              className="py-2"
+              dangerouslySetInnerHTML={{
+                __html: renderMdToHtml(announcement.content),
+              }}></div>
+          </div>
+        )}
         <>
           {session && userId === announcement.author_id ? (
             <>
@@ -83,21 +79,17 @@ export default function AnnouncementItem({
                       setIsOpen(true)
                     }}
                     className="text-gray-400 ml-2 inline-flex items-center text-sm">
-                    <>
-                      <RiDeleteBin5Line className="w-4 h-4 mr-1" />
-                      <span>delete announcement</span>
-                    </>
+                    <RiDeleteBin5Line className="w-4 h-4 mr-1" />
+                    <span>delete announcement</span>
                   </button>
                 </>
               ) : (
-                <>
-                  <button
-                    onClick={() => setEditing(true)}
-                    className="text-gray-400 mr-2 inline-flex items-center text-sm">
-                    <span>edit</span>
-                    <FaEdit className="w-4 h-4 ml-1" />
-                  </button>
-                </>
+                <button
+                  onClick={() => setEditing(true)}
+                  className="text-gray-400 mr-2 inline-flex items-center text-sm">
+                  <span>edit</span>
+                  <FaEdit className="w-4 h-4 ml-1" />
+                </button>
               )}
             </>
           ) : (
@@ -105,16 +97,13 @@ export default function AnnouncementItem({
           )}
         </>
       </div>
-
       <AlertDialog isOpen={isOpen} leastDestructiveRef={cancelRef} onClose={onClose}>
         <AlertDialogOverlay>
           <AlertDialogContent>
             <AlertDialogHeader fontSize="lg" fontWeight="bold">
-              Delete Post
+              Delete Announcement
             </AlertDialogHeader>
-
             <AlertDialogBody>Are you sure? You cannot undo this action afterwards.</AlertDialogBody>
-
             <AlertDialogFooter>
               <Button ref={cancelRef} onClick={onClose}>
                 Cancel
@@ -124,12 +113,11 @@ export default function AnnouncementItem({
                 onClick={() => {
                   deleteAnnouncement(announcement.id)
                   toast({
-                    title: 'Post deleted',
+                    title: 'Announcement deleted',
                     status: 'warning',
                     duration: 3000,
                     isClosable: true,
                   })
-                  router.push('/dashboard', undefined, { shallow: true })
                 }}
                 ml={3}>
                 Delete
