+++ components/module/NewQuest.tsx
@@ -1,46 +1,54 @@
 import { useToast } from '@chakra-ui/react'
-import { Form, Formik, useField } from 'formik'
+import { Field, Form, Formik, useField } from 'formik'
 import { nanoid } from 'nanoid'
 import { useSession } from 'next-auth/client'
 import React from 'react'
 import * as Yup from 'yup'
 import Auth from '../common/Auth'
-import Required from '../forms/Required'
-import { Announcement, makeAnnouncement, updateAnnouncement } from './DashboardAPI'
+import { Quest, makeQuest, updateQuest } from './ModuleAPI'
 
-const defaultAnnouncement = {
+const defaultQuest = {
   id: nanoid(),
-  author_id: 'string',
-  title: '',
-  content: '',
+  description: '',
+  type: '',
+  count: 0,
+  link: '',
   created_date: 0,
+  end_date: 0,
+  reward: {
+    exp: 10,
+    badge: '',
+  },
 }
 
-export default function NewAnnouncement({
-  label = 'Make an Announcement',
-  currentAnnouncement = defaultAnnouncement,
+export default function NewQuest({
+  label = 'Create Quest',
+  currentQuest = defaultQuest,
   setEditing = function (bool) {},
 }: {
   label?: string
-  currentAnnouncement?: Announcement
+  currentQuest?: Quest
   setEditing?: (bool: boolean) => void
 }): JSX.Element {
   //Initalizing values
   const initialValues = {
-    title: currentAnnouncement.title,
-    content: currentAnnouncement.content,
+    description: currentQuest.description,
+    type: currentQuest.type,
+    exp: currentQuest.reward.exp,
+    count: currentQuest.count,
+    end_date: currentQuest.end_date,
   }
 
   //User Session
   const [session] = useSession()
 
-  //Handling Announcement request
+  //Handling Quest request
   const handleSubmitNew = (value): void => {
     value.author = session.user?.name ? session.user.name : 'Anonymous'
-    makeAnnouncement(value)
+    makeQuest(value)
   }
   const handleSubmitUpdate = (value): void => {
-    updateAnnouncement(value, currentAnnouncement)
+    updateQuest(value, currentQuest)
   }
 
   //Toast
@@ -63,14 +71,14 @@ export default function NewAnnouncement({
     <Auth>
       <div
         className="bg-white overflow-hidden shadow-md rounded-lg dark:bg-gray-800 dark:text-gray-200"
-        data-cy="newAnnouncementForm">
+        data-cy="newQuestForm">
         <Formik
           initialValues={initialValues}
           validationSchema={Yup.object({
-            title: Yup.string().required('Please enter a title'),
+            description: Yup.string().required('Please enter a description'),
           })}
           onSubmit={(values, { setSubmitting }) => {
-            if (label === 'Make an Announcement') {
+            if (label === 'Make a Quest') {
               handleSubmitNew(values)
             } else {
               handleSubmitUpdate(values)
@@ -94,33 +102,42 @@ export default function NewAnnouncement({
                   <div className="max-w-sm mx-auto md:w-full md:mx-0">
                     <div className="inline-flex items-center space-x-4">
                       <h1 className="dark:text-gray-200 text-lg font-semibold">
-                        {label === 'Make an Announcement' ? label : 'Edit an Announcement'}
+                        {label === 'Make a Quest' ? label : 'Edit a Quest'}
                       </h1>
                     </div>
                   </div>
                 </div>
                 <div className="space-y-4 bg-white dark:bg-gray-700 dark:text-gray-200">
                   <div className="items-center w-full p-4 text-gray-500 dark:text-gray-300 flex-shrink-0 flex-col">
-                    <TitleTextInput label="Title" name="title" type="text" placeholder="Title" />
+                    <label htmlFor="description">Description</label>
+                    <Field
+                      name="description"
+                      rows={5}
+                      className=" rounded-lg border-transparent flex-1 appearance-none border border-gray-300 w-full py-2 px-4 bg-white dark:bg-gray-700 text-gray-700 dark:text-gray-100 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-purple-600 focus:border-transparent"
+                      placeholder="description"></Field>
                     <br />
-                    <ContentTextArea
-                      label="Content (optional)"
-                      name="content"
-                      rows={6}
-                      placeholder="Leave a comment"
-                    />
-                    <br />
+                    <label htmlFor="type">Type</label>
+                    <div className="flex flex-row space-x-2 text-lg">
+                      <label htmlFor="type">
+                        <Field type="radio" name="type" value="quiz" className="mr-1" />
+                        Quiz
+                      </label>
+                      <label htmlFor="type">
+                        <Field type="radio" name="type" value="post" className="mr-1" />
+                        Post
+                      </label>
+                    </div>
                   </div>
                   <div className="w-full px-4 pb-4 ml-auto text-gray-500 md:w-1/3">
                     <button
                       type="submit"
                       className="py-2 px-4 bg-blue-600 hover:bg-blue-700 focus:ring-blue-500 focus:ring-offset-blue-200 text-white w-full transition ease-in duration-200 text-center text-base font-semibold shadow-md focus:outline-none focus:ring-2 focus:ring-offset-2  rounded-lg "
                       onClick={() => {
-                        if (formik.touched.title && formik.errors.title) {
-                          showToast(formik.errors.title, 'title-error')
+                        if (formik.touched.description && formik.errors.description) {
+                          showToast(formik.errors.description, 'title-error')
                         }
                       }}>
-                      {label === 'Edit Announcement' ? 'Save' : 'Post Announcement'}
+                      {label === 'Edit Quest' ? 'Save' : 'Post Quest'}
                     </button>
                   </div>
                 </div>
@@ -130,53 +147,5 @@ export default function NewAnnouncement({
         </Formik>
       </div>
     </Auth>
-  )
-}
-
-const ContentTextArea = ({
-  label,
-  ...props
-}: {
-  label: string
-  name: string
-  rows: number
-  placeholder: string
-}): JSX.Element => {
-  const [field, meta] = useField(props)
-  return (
-    <>
-      <label htmlFor={props.name}>{label}</label>
-      <textarea
-        className="flex rounded-lg border-transparent appearance-none border border-gray-300 w-full py-2 px-4 bg-white text-gray-700 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-purple-600 focus:border-transparent"
-        {...field}
-        {...props}
-      />
-      {meta.touched && meta.error ? <div className="">{meta.error}</div> : null}
-    </>
-  )
-}
-
-const TitleTextInput = ({
-  label,
-  ...props
-}: {
-  label: string
-  name: string
-  type: string
-  placeholder: string
-}): JSX.Element => {
-  const [field, meta] = useField(props)
-  return (
-    <>
-      <div className="flex items-center">
-        <label htmlFor={props.name}>{label}</label>
-        {meta.touched && meta.error ? <Required /> : null}
-      </div>
-      <input
-        className="flex rounded-lg border-transparent appearance-none border border-gray-300 w-full py-2 px-4 bg-white text-gray-700 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-purple-600 focus:border-transparent"
-        {...field}
-        {...props}
-      />
-    </>
   )
 }
