+++ pages/module/index.tsx
@@ -1,100 +1,88 @@
-import { useToast } from '@chakra-ui/react'
-import { ErrorMessage, Field, Form, Formik } from 'formik'
-import { nanoid } from 'nanoid'
+import { Button } from '@chakra-ui/react'
 import Head from 'next/head'
-import * as Yup from 'yup'
-import Layout from '../../components/layouts/Layout'
-import { makeModule } from '../../components/module/ModuleAPI'
+import { useRouter } from 'next/router'
+import React from 'react'
+import SidebarLayout from '../../components/layouts/SidebarLayout'
+import { useAllModules } from '../../components/module/ModuleAPI'
+
 export default function Home(): JSX.Element {
-  const initialValues = {
-    title: 'CS2030',
-    users: [
-      {
-        id: 'ddHg168Fwz9VIP1wxbzK',
-        name: 'Dreamer',
-        exp: 0,
-        role: 'admin',
-        badges: ['a', 'b'],
-        completedTasks: [],
-      },
-    ],
-    questions: [],
-    quiz: [],
-    forum: [],
-    reply: [],
-    tasks: [],
-    schedule: [],
-  }
-  const handleSubmit = (value): void => {
-    value.id = nanoid()
-    console.log(value)
-    makeModule(value)
-  }
-  const toast = useToast()
+  const router = useRouter()
+  const { modules, isLoading } = useAllModules()
   return (
     <>
       <Head>
-        <title>Create Module | NUS Connect</title>
+        <title>Modules | NUS Connect</title>
         <meta name="description" content="NUS Connect" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Layout>
-        <Formik
-          initialValues={initialValues}
-          validationSchema={Yup.object({
-            title: Yup.string().required('Required'),
-          })}
-          onSubmit={(values, { setSubmitting, resetForm }) => {
-            handleSubmit(values)
-            setTimeout(() => {
-              toast({
-                title: 'Done!',
-                status: 'success',
-                duration: 5000,
-                isClosable: true,
-                position: 'top-right',
-              })
-              setSubmitting(false)
-              resetForm()
-            }, 400)
-          }}>
-          {(formik) => (
-            <section className="bg-gray-100 bg-opacity-20">
-              <Form className="container max-w-3xl mx-auto shadow-md md:w-3/4">
-                <div className="p-4 bg-gray-100 border-t-2 border-indigo-400 rounded-lg bg-opacity-5">
-                  <div className="max-w-sm mx-auto md:w-full md:mx-0">
-                    <div className="inline-flex items-center space-x-4">
-                      <h1 className="text-gray-600">Create a module</h1>
-                    </div>
-                  </div>
-                </div>
-                <div className="space-y-6 bg-white">
-                  <div className="items-center w-full p-4 space-y-4 text-gray-500 md:inline-flex md:space-y-0">
-                    <h2 className="max-w-sm mx-auto md:w-2/12">Module Title</h2>
-                    <div className="max-w-md mx-auto md:w-10/12">
-                      <div className=" relative ">
-                        <Field
-                          name="title"
-                          rows={5}
-                          className=" rounded-lg border-transparent flex-1 appearance-none border border-gray-300 w-full py-2 px-4 bg-white text-gray-700 placeholder-gray-400 shadow-sm text-base focus:outline-none focus:ring-2 focus:ring-purple-600 focus:border-transparent"
-                          placeholder="title"></Field>
-                        <ErrorMessage name="title" />
-                      </div>
-                    </div>
-                  </div>
-                  <div className="w-full px-4 pb-4 ml-auto text-gray-500 md:w-1/3">
-                    <button
-                      type="submit"
-                      className="py-2 px-4  bg-blue-600 hover:bg-blue-700 focus:ring-blue-500 focus:ring-offset-blue-200 text-white w-full transition ease-in duration-200 text-center text-base font-semibold shadow-md focus:outline-none focus:ring-2 focus:ring-offset-2  rounded-lg ">
-                      Save
-                    </button>
-                  </div>
-                </div>
-              </Form>
-            </section>
-          )}
-        </Formik>
-      </Layout>
+      <SidebarLayout>
+        <div className="pt-12 md:pt-2 relative flex flex-col min-w-0 mb-4 lg:mb-0 break-words bg-gray-50 dark:bg-gray-800 w-full rounded">
+          <div className="rounded-t mb-0 px-0 border-0">
+            <div className="flex flex-wrap items-center px-4 py-2">
+              <div className="relative w-full max-w-full flex-grow flex-1">
+                <h3 className="font-semibold text-base text-gray-900 dark:text-gray-50">
+                  Existing Modules
+                </h3>
+              </div>
+              <div className="relative w-full max-w-full flex-grow flex-1 text-right">
+                <button
+                  className="bg-blue-500 dark:bg-gray-100 text-white active:bg-blue-600 dark:text-gray-800 dark:active:text-gray-700 text-xs font-bold uppercase px-3 py-1 rounded outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                  onClick={() => router.push('module/new-module')}
+                  type="button">
+                  Want to create a module ?
+                </button>
+              </div>
+            </div>
+            <div className="block w-full overflow-x-auto">
+              <table className="items-center w-full bg-transparent border-collapse">
+                <thead>
+                  <tr>
+                    <th className="px-4 bg-gray-100 dark:bg-gray-600 text-gray-500 dark:text-gray-100 align-middle border border-solid border-gray-200 dark:border-gray-500 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left">
+                      Module
+                    </th>
+                    <th className="px-4 bg-gray-100 dark:bg-gray-600 text-gray-500 dark:text-gray-100 align-middle border border-solid border-gray-200 dark:border-gray-500 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left">
+                      Semester
+                    </th>
+                    <th className="px-4 bg-gray-100 dark:bg-gray-600 text-gray-500 dark:text-gray-100 align-middle border border-solid border-gray-200 dark:border-gray-500 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left">
+                      Professor
+                    </th>
+                    <th className="px-4 bg-gray-100 dark:bg-gray-600 text-gray-500 dark:text-gray-100 align-middle border border-solid border-gray-200 dark:border-gray-500 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left min-w-140-px">
+                      Number of Students
+                    </th>
+                    <th className="px-4 bg-gray-100 dark:bg-gray-600 text-gray-500 dark:text-gray-100 align-middle border border-solid border-gray-200 dark:border-gray-500 py-3 text-xs uppercase border-l-0 border-r-0 whitespace-nowrap font-semibold text-left min-w-140-px">
+                      Status
+                    </th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {!isLoading &&
+                    modules.map((module) => (
+                      <tr className="text-gray-700 dark:text-gray-100" key={module.id}>
+                        <th className="border-t-0 px-4 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4 text-left">
+                          {module.title}
+                        </th>
+                        <td className="border-t-0 px-4 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4">
+                          2021/2022 Sem 1
+                        </td>
+                        <td className="border-t-0 px-4 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4">
+                          Prof X
+                        </td>
+                        <td className="border-t-0 px-4 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4">
+                          <div className="flex items-center">
+                            <span className="mr-2">{module.users.length}</span>
+                          </div>
+                        </td>
+                        <td className="border-t-0 px-4 align-middle border-l-0 border-r-0 text-xs whitespace-nowrap p-4">
+                          <Button className="dark:text-black">Join</Button>
+                        </td>
+                      </tr>
+                    ))}
+                </tbody>
+              </table>
+            </div>
+          </div>
+        </div>
+      </SidebarLayout>
     </>
   )
 }
