--- public/data/IndividualModuleData.json
@@ -1,33 +1,0 @@
-[{
-    "id": "xft5nj9NXr_RXl3LEyt2g",
-    "title": "CS2030",
-    "semester": "2021/2022 Sem 1",
-    "professor": "Prof Henry",
-    "user": "level 2"
-
-  },
-  {
-    "id": "0s7Ej13iFqWCtDhson5mx",
-    "title": "CS2040",
-    "semester": "2021/2022 Sem 2",
-    "professor": "Prof Ben",
-    "user": "level 3"
-
-  },
-  {
-    "id": "bYUZiLoC45FU_IvRSQgJZ",
-    "title": "CS2103T",
-    "semester": "2021/2022 Sem 1",
-    "professor": "Prof Damith",
-    "user": "level 4"
-
-  },
-  {
-    "id": "fsYWvAe8LJfLKQgmraByL",
-    "title": "CS1010",
-    "semester": "2021/2022 Sem 2",
-    "professor": "Prof Bobo",
-    "user": "level 5"
-
-  }
-]
