--- public/data/ModuleDataNew.json
@@ -1,50 +1,0 @@
-[{
-  "id": "xft5nj9NXr_RXl3LEyt2g",
-  "title": "CS2030",
-  "schedule": [{
-    "id": "rOS6Pt4tnuaSYrbZ2hpQF",
-    "week": "1",
-    "announcements": [{
-        "id": "id1",
-        "author_id": "WnqaRimjGDDY9KxM49XQ",
-        "title": "Welcome to your first week of CS2030!",
-        "content": "Some things to note for this week ...",
-        "created_date": 1622557680592
-      },
-      {
-        "id": "id2",
-        "author_id": "WnqaRimjGDDY9KxM49XQ",
-        "title": "Welcome to your second week of CS2030!",
-        "content": "Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...Some things to note for this week ...",
-        "created_date": 1622557680592
-      }
-    ],
-    "quests": [{
-        "id": "MpGV4bzRpbNj683Pt-ful",
-        "description": "Do a quiz",
-        "type": "quiz",
-        "count": 1,
-        "link": "",
-        "created_date": 1622557680592,
-        "end_date": 1622557680592,
-        "reward": {
-          "exp": 30,
-          "badge": "new badge"
-        }
-      },
-      {
-        "id": "test2",
-        "description": "Post a forum post",
-        "type": "forum",
-        "count": 1,
-        "link": "",
-        "created_date": 1622557680592,
-        "end_date": 1622557680592,
-        "reward": {
-          "exp": 20,
-          "badge": "new badge"
-        }
-      }
-    ]
-  }]
-}]
